Assembler.nw: mk sync from principia with many fixes
long->int32, ncu fix, stmtline fix

+++ assemblers/as/float.c
@@ -29,7 +29,7 @@ ieeedtod(Ieee *ieee, double native)
     fr = modf(fr*f, &ho);
     ieee->l = ho;
     ieee->l <<= 16;
-    ieee->l |= (long)(fr*f);
+    ieee->l |= (int32)(fr*f);
 }
 /*e: function [[ieeedtod]] */
